requirements.txt
Faker==37.6.0
matplotlib==3.10.5
pandas==2.3.0
plotly==6.1.2
pyarrow==20.0.0
scikit-learn==1.7.1
streamlit-option-menu
xgboost==3.0.5
